move cache-related functions to arc.ts
